chore: check sus nodes periodically and not on query
