+++ sn/src/node/api/dispatcher.rs
@@ -186,9 +186,29 @@ impl Dispatcher {
                         .enqueue_and_handle_next_cmd_and_offshoots(cmd, None)
                         .await
                     {
-                        error!("Error sending a cleaning up unused PeerLinks: {:?}", e);
+                        error!(
+                            "Error sending Propose Offline for dysfunctional nodes: {:?}",
+                            e
+                        );
                     }
                 }
+
+                match dispatcher.node.notify_about_newly_suspect_nodes().await {
+                    Ok(suspect_cmds) => {
+                        for cmd in suspect_cmds {
+                            if let Err(e) = dispatcher
+                                .clone()
+                                .enqueue_and_handle_next_cmd_and_offshoots(cmd, None)
+                                .await
+                            {
+                                error!("Error processing suspect node cmds: {:?}", e);
+                            }
+                        }
+                    }
+                    Err(error) => {
+                        error!("Error getting suspect nodes: {error}");
+                    }
+                };
             }
         });
     }
